+++ lavis/models/sevila_models/sevila.py
@@ -35,7 +35,7 @@ class SeViLA(Blip2Base):
     }
 
     def __init__( self, img_size=224, drop_path_rate=0,
-        use_grad_checkpoint=False, vit_precision="fp16", freeze_vit=True,
+        use_grad_checkpoint=False, vit_precision="fp32", freeze_vit=True,
         num_query_token=32, t5_model="google/flan-t5-xl", prompt="",
         max_txt_len=32, frame_num=8, answer_num=5, apply_lemmatizer=False, task='qa'):
         """
@@ -157,7 +157,7 @@ class SeViLA(Blip2Base):
                 answer = samples['qa_output']
                 ans_idx = [self.ANS_MAP[a[-1]] for a in answer]
 
-                with torch.cuda.amp.autocast(dtype=torch.bfloat16):
+                with torch.cuda.amp.autocast(dtype=torch.float32):
                     # Frame Prefix
                     frame_prefix = self.t5_tokenizer(
                         self.frame_prefix, padding="longest", add_special_tokens=False,
@@ -214,7 +214,7 @@ class SeViLA(Blip2Base):
                 encoder_attention_mask=image_atts, return_dict=True) # bt, n, c
             inputs_t5_loc = self.t5_proj_loc(query_output_loc.last_hidden_state) # bt, n, c
             atts_t5_loc = torch.ones(inputs_t5_loc.size()[:-1], dtype=torch.long).to(image.device)
-            with torch.cuda.amp.autocast(dtype=torch.bfloat16):
+            with torch.cuda.amp.autocast(dtype=torch.float32):
                 frame_prefix = self.t5_tokenizer(
                     self.frame_prefix, padding="longest", add_special_tokens=False,
                     truncation=True, max_length=self.max_txt_len, return_tensors="pt").to(image.device) 
@@ -263,7 +263,7 @@ class SeViLA(Blip2Base):
                 inputs_t5_loc = self.t5_proj_loc(query_output_loc.last_hidden_state)
 
                 atts_t5_loc = torch.ones(inputs_t5_loc.size()[:-1], dtype=torch.long).to(image.device)
-                with torch.cuda.amp.autocast(dtype=torch.bfloat16):
+                with torch.cuda.amp.autocast(dtype=torch.float32):
 
                     frame_prefix = self.t5_tokenizer(
                         self.frame_prefix, padding="longest", add_special_tokens=False,
@@ -318,7 +318,7 @@ class SeViLA(Blip2Base):
             inputs_t5_qa = inputs_t5_qa.reshape(b, -1, inputs_t5_qa.shape[-2], inputs_t5_qa.shape[-1])
             atts_t5_qa = torch.ones(inputs_t5_qa.size()[:-1], dtype=torch.long).to(image.device)
             
-            with torch.cuda.amp.autocast(dtype=torch.bfloat16):        
+            with torch.cuda.amp.autocast(dtype=torch.float32):        
                 vid_prefix = self.t5_tokenizer(
                     self.vid_prefix, padding="longest", add_special_tokens=False,
                     truncation=True, max_length=self.max_txt_len, return_tensors="pt",).to(image.device) # 
@@ -365,7 +365,7 @@ class SeViLA(Blip2Base):
             text_input_qa = samples['qa_input'] 
             answer = samples['qa_output'] 
 
-            with torch.cuda.amp.autocast(dtype=torch.bfloat16):
+            with torch.cuda.amp.autocast(dtype=torch.float32):
                     # Frame Prefix
                 if 'qa_vid' not in self.task:
                     atts_t5_qa = torch.ones(inputs_t5_qa.size()[:-1], dtype=torch.long).to(image.device) 
@@ -475,7 +475,7 @@ class SeViLA(Blip2Base):
                 encoder_attention_mask=image_atts, return_dict=True)
             inputs_t5_qa = self.t5_proj(query_output_qa.last_hidden_state)
             
-            with torch.cuda.amp.autocast(dtype=torch.bfloat16):
+            with torch.cuda.amp.autocast(dtype=torch.float32):
                     # Frame Prefix
                 if 'vid' not in self.task: 
                     atts_t5_qa = torch.ones(inputs_t5_qa.size()[:-1], dtype=torch.long).to(image.device) 
@@ -554,7 +554,7 @@ class SeViLA(Blip2Base):
             inputs_t5_loc = self.t5_proj_loc(query_output_loc.last_hidden_state)
 
             atts_t5_loc = torch.ones(inputs_t5_loc.size()[:-1], dtype=torch.long).to(image.device)
-            with torch.cuda.amp.autocast(dtype=torch.bfloat16):
+            with torch.cuda.amp.autocast(dtype=torch.float32):
 
                 frame_prefix = self.t5_tokenizer(
                     self.frame_prefix, padding="longest", add_special_tokens=False,
@@ -736,7 +736,7 @@ class SeViLA(Blip2Base):
         inputs_t5_loc = self.t5_proj_loc(query_output_loc.last_hidden_state)
 
         atts_t5_loc = torch.ones(inputs_t5_loc.size()[:-1], dtype=torch.long).to(image.device)
-        with torch.cuda.amp.autocast(dtype=torch.bfloat16):
+        with torch.cuda.amp.autocast(dtype=torch.float32):
 
             frame_prefix = self.t5_tokenizer(
                 self.frame_prefix, padding="longest", add_special_tokens=False,
@@ -913,7 +913,7 @@ class SeViLA(Blip2Base):
         encoder_atts = torch.cat([atts_t5, input_tokens.attention_mask], dim=1)
 
         device_type = "cuda" if "cuda" in str(self.device) else "cpu"
-        with torch.amp.autocast(device_type=device_type, dtype=torch.bfloat16):
+        with torch.amp.autocast(device_type=device_type, dtype=torch.float32):
             inputs_embeds = self.t5_model.encoder.embed_tokens(input_tokens.input_ids)
             inputs_embeds = torch.cat([inputs_t5, inputs_embeds], dim=1)
